refactor(cloudrun): share handler error classes + dispatcher mapper (C6) (#70)
- InvalidArgsError / IdentityRequiredError / UnauthorizedError move to
  @tokens/cloudrun-shutdown/http-errors (subpath export). Previously
  defined identically in cloudrun-admin, cloudrun-usage, and twice in
  cloudrun-assets. Each app's handlers/errors.ts becomes a re-export so
  the ~20 importers are untouched.
- The byte-identical (minus service prefix) errorResponse dispatcher
  mapping in admin/usage collapses into a shared
  dispatchErrorResponse(serviceName, ...).
- cloudrun-assets' duplicated InvalidArgsError definitions (assets.ts +
  crons.ts) re-export the shared class — server.ts's
  InvalidArgsError-vs-CronInvalidArgsError disambiguation now resolves
  to one class (instanceof behavior unchanged).

650 tests green across the three services.
